requirements.txt
Flask
pandas
numpy
scikit-learn
nltk
requests
python-dotenv
gunicorn

